+++ trunk/SP/code/qcommon/vm.c
@@ -816,7 +816,7 @@ intptr_t QDECL VM_Call( vm_t *vm, intptr_t callnum, ... )
 	lastVM = vm;
 
 	if ( vm_debugLevel ) {
-	  Com_Printf( "VM_Call( %d )\n", callnum );
+	  Com_Printf( "VM_Call( %d )\n", (int)callnum );
 	}
 
 	++vm->callLevel;
