+++ trunk/SP/code/sys/sys_main.c
@@ -478,7 +478,7 @@ Used to load a development dll instead of a virtual machine
 =================
 */
 void *Sys_LoadGameDll(const char *name,
-	intptr_t (QDECL **entryPoint)(int, ...),
+	intptr_t (QDECL **entryPoint)(intptr_t, ...),
 	intptr_t (*systemcalls)(intptr_t, ...))
 {
 	void *libHandle;
